+++ inc/lib/admin.php
@@ -6,7 +6,7 @@
  * @copyright  2015 WebMan - Oliver Juhas
  *
  * @since    1.0
- * @version  1.4.2
+ * @version  1.4.5
  *
  * CONTENT:
  * - 10) Actions and filters
@@ -51,7 +51,7 @@
 	 * Admin HTML head assets enqueue
 	 *
 	 * @since    1.0
-	 * @version  1.4.2
+	 * @version  1.4.5
 	 */
 	if ( ! function_exists( 'wm_assets_admin' ) ) {
 		function wm_assets_admin() {
@@ -71,12 +71,6 @@
 								'screen'
 							);
 
-					//Styles - inline
-						$custom_styles .= '#post-' . get_option( 'page_on_front' ) . ' { background: #d7eef4; }';
-						$custom_styles .= '#post-' . get_option( 'page_for_posts' ) . ' { background: #d7f4e3; }';
-
-						wp_add_inline_style( 'wm-admin-styles', apply_filters( 'wmhook_esc_css', $custom_styles ) );
-
 				}
 		}
 	} // /wm_assets_admin
@@ -93,7 +87,7 @@
 	 * Register table columns
 	 *
 	 * @since    1.0
-	 * @version  1.0
+	 * @version  1.4.5
 	 *
 	 * @param  array $columns
 	 */
@@ -104,11 +98,11 @@
 					unset( $columns['thumbnail'] );
 				}
 
-				$add             = array_slice( $columns, 0, 1 );
+				$add             = array_slice( $columns, 0, 2 );
 				$add['wm-thumb'] = __( 'Image', 'wm_domain' );
 
 			//Output
-				return apply_filters( 'wmhook_wm_post_columns_register_output', array_merge( $add, array_slice( $columns, 1 ) ) );
+				return apply_filters( 'wmhook_wm_post_columns_register_output', array_merge( $add, array_slice( $columns, 2 ) ) );
 		}
 	} // /wm_post_columns_register