+++ inc/setup.php
@@ -6,7 +6,7 @@
  * @copyright  2015 WebMan - Oliver Juhas
  *
  * @since    1.0
- * @version  1.4
+ * @version  1.4.5
  *
  * CONTENT:
  * -  10) Actions and filters
@@ -666,7 +666,7 @@
 	 * Registering theme styles and scripts
 	 *
 	 * @since    1.0
-	 * @version  1.4.2
+	 * @version  1.4.5
 	 */
 	if ( ! function_exists( 'wm_register_assets' ) ) {
 		function wm_register_assets() {
@@ -703,7 +703,8 @@
 				$register_scripts = apply_filters( 'wmhook_wm_register_assets_register_scripts', array(
 						'wm-imagesloaded'        => array( wm_get_stylesheet_directory_uri( 'js/imagesloaded.pkgd.min.js' ) ),
 						'wm-slick'               => array( wm_get_stylesheet_directory_uri( 'js/slick.min.js' ) ),
-						'wm-scripts-global'      => array( 'src' => wm_get_stylesheet_directory_uri( 'js/scripts-global.js' ), 'deps' => array( 'wm-imagesloaded', 'wm-slick' ) ),
+						'wm-scripts-global'      => array( 'src' => wm_get_stylesheet_directory_uri( 'js/scripts-global.js' ), 'deps' => array( 'wm-imagesloaded', 'wm-slick', 'wm-scripts-navigation' ) ),
+						'wm-scripts-navigation'  => array( wm_get_stylesheet_directory_uri( 'js/scripts-navigation.js' ) ),
 						'wm-skip-link-focus-fix' => array( wm_get_stylesheet_directory_uri( 'js/skip-link-focus-fix.js' ) ),
 					) );
 
@@ -1221,20 +1222,48 @@
 
 		/**
 		 * Navigation item improvements
+		 *
+		 * @since    1.0
+		 * @version  1.4.5
 		 */
 		if ( ! function_exists( 'wm_nav_item_process' ) ) {
 			function wm_nav_item_process( $item_output, $item, $depth, $args ) {
-				//Preparing output
-					//Display item description
-						if (
-								'primary' == $args->theme_location
-								&& trim( $item->description )
-							) {
-							$item_output = str_replace( $args->link_after . '</a>', '<span class="menu-item-description">' . trim( $item->description ) . '</span>' . $args->link_after . '</a>', $item_output );
-						}
 
-				//Output
+				// Processing
+
+					if ( 'primary' == $args->theme_location ) {
+
+						// Description
+
+							if ( trim( $item->description ) ) {
+
+								$item_output = str_replace(
+										$args->link_after . '</a>',
+										'<span class="menu-item-description">' . trim( $item->description ) . '</span>' . $args->link_after . '</a>',
+										$item_output
+									);
+
+							}
+
+						// Submenu expander button
+
+							if ( in_array( 'menu-item-has-children', (array) $item->classes ) ) {
+
+								$item_output = str_replace(
+										$args->link_after . '</a>',
+										$args->link_after . ' <span class="expander"></span></a>',
+										$item_output
+									);
+
+							}
+
+					}
+
+
+				// Output
+
 					return $item_output;
+
 			}
 		} // /wm_nav_item_process
 
@@ -1506,7 +1535,7 @@
 		 * Post thumbnail (featured image) display size
 		 *
 		 * @since    1.3
-		 * @version  1.3
+		 * @version  1.4.5
 		 *
 		 * @param  string $image_size
 		 */
@@ -1515,7 +1544,7 @@
 				//Preparing output
 					if (
 							is_single( get_the_ID() )
-							|| is_page()
+							|| is_page( get_the_ID() )
 						) {
 
 						$image_size = 'large';